+++ lite/iarg.cpp
@@ -107,7 +107,7 @@ val_.str_ = str;
 #endif
 }
 
-Iarg::Iarg(long num)
+Iarg::Iarg(long long num)
 {
 type = IANUM;
 val_.num_ = num;
@@ -634,14 +634,14 @@ if (type != IASTR)
 return val_.str_;
 }
 
-long Iarg::getNum()
+long long Iarg::getNum()
 {
 if (type != IANUM)
 	{
 	std::_t_strstream gerrStr;
 	gerrStr << _T("[getNum: Arg is not a num.]") << std::ends;
 	errOut(&gerrStr,false);
-	return 0;
+	return 0LL;
 	}
 return val_.num_;
 }
@@ -689,7 +689,7 @@ return val_.sem_;
 }
 
 // FOR ODBC BINDING.															// 06/08/02 AM.
-long *Iarg::getNumaddr()													// 06/08/02 AM.
+long long *Iarg::getNumaddr()													// 06/08/02 AM.
 {
 return &(val_.num_);
 }
@@ -705,7 +705,7 @@ return &(val_.float_);
 ********************************************/
 
 void Iarg::setType(enum Iargtype x)	{ type = x; }
-void Iarg::setNum(long x)
+void Iarg::setNum(long long x)
 {
 if (type != IANUM)
 	{
@@ -781,7 +781,7 @@ val_.sem_ = x;
 * OPT:	Would like semantics to stop storing numeric args as strings.
 ********************************************/
 
-bool Iarg::fetchNum(/*UP*/ long &num)
+bool Iarg::fetchNum(/*UP*/ long long &num)
 {
 num = -1;
 switch (type)
